feat: 完善寝室和成员管理crud逻辑；修复windows下的nest重启问题（win下的redis版本太低）

+++ src/modules/class/class.service.ts
@@ -3,7 +3,9 @@ import { CreateDormitoryDto, CreateMemberDto } from './dto/create-class.dto';
 import { UpdateMemberDto, UpdateRoomDto } from './dto/update-class.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ClassPhotos, Dormitory } from './entities/class.entity';
-import { Repository } from 'typeorm';
+import { Like, Repository, getRepository } from 'typeorm';
+import { isArray } from 'class-validator';
+import { QueryClassDto } from './dto/query-class.dto';
 
 @Injectable()
 export class ClassService {
@@ -31,7 +33,13 @@ export class ClassService {
       .getMany();
     for (const dormitory of res) {
       dormitory['photo_nums'] = dormitory['members'].length;
-      dormitory['swiper_list'] = dormitory['swiper_list'].split('\\n');
+      for (const dormitory of res) {
+        dormitory['photo_nums'] = dormitory['members'].length;
+        dormitory['members'].forEach((member) => {
+          member['dormitoryId'] = dormitory.dormitoryId;
+        });
+      }
+      // dormitory['swiper_list'] = dormitory['swiper_list'].split('\\n');
     }
     return res;
   }
@@ -48,20 +56,23 @@ export class ClassService {
     return await this.dormitoryRepo.save(createDormitoryDto);
   }
 
-  // 更新寝室信息
+  // 更新寝室信息(还是要根据主键去更新)
   async updateRoom(updateClassDto: UpdateRoomDto): Promise<any> {
     console.log('updateClassDto', updateClassDto);
     const roomInfo = await this.dormitoryRepo.findOne({
-      where: { dormitoryId: updateClassDto.dormitoryId },
+      where: { id: updateClassDto.id },
     });
     if (!roomInfo) throw new BadRequestException('寝室信息不存在或者已删除');
+    if (isArray(updateClassDto.swiper_list)) {
+      updateClassDto.swiper_list = updateClassDto.swiper_list.join('\\n');
+    }
     const newClass = this.dormitoryRepo.merge(roomInfo, updateClassDto);
     return await this.dormitoryRepo.save(newClass);
   }
 
-  // 删除寝室
+  // 删除寝室（修改成主键删除即可）
   async removeRoom(id: number) {
-    const isExist = await this.dormitoryRepo.exist({ where: { dormitoryId: id } });
+    const isExist = await this.dormitoryRepo.exist({ where: { id } });
     if (!isExist) {
       throw new BadRequestException('寝室不存在或者已删除');
     }
@@ -70,13 +81,26 @@ export class ClassService {
 
   // 创建成员信息
   async createMember(createMemberDto: CreateMemberDto) {
+    // 寝室表查询是否存在
     const isExist = await this.dormitoryRepo.findOne({
       where: { dormitoryId: createMemberDto.dormitoryId },
     });
     if (!isExist) {
       throw new BadRequestException('寝室不存在');
     }
-    return await this.classRepo.save(createMemberDto);
+    // console.log('isExist', isExist);
+
+    // 创建 ClassPhotos 实例
+    const classPhotos = new ClassPhotos();
+    classPhotos.name = createMemberDto.name || '';
+    classPhotos.imageUrl = createMemberDto.imageUrl;
+    const dormitoryId = createMemberDto.dormitoryId; // 假设宿舍的 dormitoryId 为 1
+    const dormitory = await this.dormitoryRepo.findOne({ where: { dormitoryId } });
+    // 设置 ClassPhotos 实例的 dormitory 属性
+    classPhotos.dormitory = dormitory;
+
+    // 保存 ClassPhotos 实例到数据库中
+    return await this.classRepo.save(classPhotos);
   }
 
   // 更新单个同学信息
@@ -96,6 +120,10 @@ export class ClassService {
   }
   // 删除单个同学
   async removeMember(id: number) {
+    const isExist = await this.classRepo.exist({ where: { id } });
+    if (!isExist) {
+      throw new BadRequestException('成员不存在或者已删除');
+    }
     return await this.classRepo.delete(id);
   }
 
@@ -103,5 +131,56 @@ export class ClassService {
   async getAllRomIDs() {
     const res = await this.dormitoryRepo.find({ select: ['id', 'dormitoryId', 'title', 'desc'] });
     return res;
+  }
+  // 分页查询带参数
+  async findRoomInfoPagination(query: QueryClassDto) {
+    const pageSize = query.pageSize || 10;
+    const pageNo = query.pageNo || 1;
+    const [data, total] = await this.dormitoryRepo.findAndCount({
+      where: {
+        title: Like(`%${query.title || ''}%`),
+        dormitoryId: query.dormitoryId,
+      },
+      relations: { members: true },
+      order: {
+        id: 'ASC',
+      },
+      take: pageSize,
+      skip: (pageNo - 1) * pageSize,
+    });
+    const pageData = data.map((item) => {
+      const permissionIds = item.members.map((p) => p.id);
+      item.swiper_list = item.swiper_list.split('\\n');
+      item.photo_nums = item.members.length;
+      delete item.members;
+      return { ...item, permissionIds };
+    });
+    return { pageData, total };
+  }
+  // 分页查询带参数（成员信息）
+  async findMemberInfoPagination(query: QueryClassDto) {
+    const pageSize = query.pageSize || 10;
+    const pageNo = query.pageNo || 1;
+    const [data, total] = await this.classRepo.findAndCount({
+      where: {
+        name: Like(`%${query.name || ''}%`),
+        dormitory: {
+          dormitoryId: query.dormitoryId,
+        },
+      },
+      relations: ['dormitory'],
+      order: {
+        id: 'ASC',
+      },
+      take: pageSize,
+      skip: (pageNo - 1) * pageSize,
+    });
+    const pageData = data.map((item) => {
+      const dormitoryId = item.dormitory.dormitoryId;
+      // item.swiper_list = item.swiper_list.split('\\n');
+      delete item.dormitory;
+      return { ...item, dormitoryId };
+    });
+    return { pageData, total };
   }
 }
